Gross Commit
Lots of changes. Currently it does show stuff, so im committing while it "works"

I kind of modelled our site after some example page I saw where they had their logo and some text on the right side of the page. I ran with that and put our site name on it and a little welcome banner.

Added nav bar at the top. Not sure why it only spans like 40% of the screen. I have to go through and try to fix that, or change the nav bar color to match the background so it looks intentional that it doesnt cover the full screen.

I may have randomly/accidentally moved files. sorry about that.